+++ spec/legion/extensions/mesh/transport/queues/gossip_spec.rb
@@ -12,17 +12,25 @@ unless defined?(Legion::Transport::Queue)
   end
 end
 
-unless defined?(Legion::Transport::Exchanges::Node)
+unless defined?(Legion::Transport::Queue)
   module Legion
     module Transport
-      module Exchanges
-        class Node # rubocop:disable Lint/EmptyClass
-        end
+      class Queue
+        def initialize(**); end
       end
     end
   end
 end
 
+unless defined?(Legion::Transport::Exchange)
+  module Legion
+    module Transport
+      class Exchange; end # rubocop:disable Lint/EmptyClass
+    end
+  end
+end
+
+require 'legion/extensions/mesh/transport/exchanges/fanout'
 require 'legion/extensions/mesh/transport/queues/gossip'
 
 RSpec.describe Legion::Extensions::Mesh::Transport::Queues::Gossip do
@@ -35,14 +43,8 @@ RSpec.describe Legion::Extensions::Mesh::Transport::Queues::Gossip do
   end
 
   describe '#exchange' do
-    it 'returns Legion::Transport::Exchanges::Node' do
-      expect(queue.exchange).to eq(Legion::Transport::Exchanges::Node)
-    end
-  end
-
-  describe '#routing_key' do
-    it "returns 'mesh.gossip'" do
-      expect(queue.routing_key).to eq('mesh.gossip')
+    it 'returns the Fanout exchange' do
+      expect(queue.exchange).to eq(Legion::Extensions::Mesh::Transport::Exchanges::Fanout)
     end
   end
 end